refactor: extract shared TTL options, time utils, and isExpired helper
- client/src/utils/ttl.ts: shared TTL_OPTIONS + TTL_OPTIONS_WITH_NONE (adds "No expiry" for ScheduleModal)
- client/src/utils/time.ts: shared relativeTime() and expiresIn()
- FanOutModal + ScheduleModal: import from shared utils, remove local duplicates
- cronService: isExpired() helper replaces 3 inline expiry checks; pre-parse expiresAt to ms at job registration so the cron callback compares integers instead of constructing Date objects on every tick; startup log only mentions discarded count when >0
