+++ client/src/utils/time.ts
@@ -1,0 +1,18 @@
+export function relativeTime(iso: string): string {
+  const mins = Math.floor((Date.now() - new Date(iso).getTime()) / 60000);
+  if (mins < 1) return 'just now';
+  if (mins < 60) return `${mins}m ago`;
+  const hrs = Math.floor(mins / 60);
+  if (hrs < 24) return `${hrs}h ago`;
+  return `${Math.floor(hrs / 24)}d ago`;
+}
+
+export function expiresIn(iso: string): string {
+  const diff = new Date(iso).getTime() - Date.now();
+  if (diff <= 0) return 'expired';
+  const mins = Math.floor(diff / 60000);
+  if (mins < 60) return `expires in ${mins}m`;
+  const hrs = Math.floor(mins / 60);
+  if (hrs < 24) return `expires in ${hrs}h`;
+  return `expires in ${Math.floor(hrs / 24)}d`;
+}
